+++ src/clone-node.ts
@@ -125,6 +125,7 @@ function cloneCSSStyle<T extends HTMLElement>(nativeNode: T, clonedNode: T) {
           Math.floor(parseFloat(value.substring(0, value.length - 2))) - 0.1
         value = `${reducedFont}px`
       }
+
       if (
         isInstanceOfElement(nativeNode, HTMLIFrameElement) &&
         name === 'display' &&
@@ -132,6 +133,11 @@ function cloneCSSStyle<T extends HTMLElement>(nativeNode: T, clonedNode: T) {
       ) {
         value = 'block'
       }
+      
+      if (name === 'd' && clonedNode.getAttribute('d')) {
+        value = `path(${clonedNode.getAttribute('d')})`
+      }
+      
       targetStyle.setProperty(
         name,
         value,