+++ src/embed-images.ts
@@ -4,18 +4,33 @@ import { toArray, isInstanceOfElement } from './util'
 import { isDataUrl, resourceToDataURL } from './dataurl'
 import { getMimeType } from './mimes'
 
+async function embedProp(
+  propName: string,
+  node: HTMLElement,
+  options: Options,
+) {
+  const propValue = node.style?.getPropertyValue(propName)
+  if (propValue) {
+    const cssString = await embedResources(propValue, null, options)
+    node.style.setProperty(
+      propName,
+      cssString,
+      node.style.getPropertyPriority(propName),
+    )
+    return true
+  }
+  return false
+}
+
 async function embedBackground<T extends HTMLElement>(
   clonedNode: T,
   options: Options,
 ) {
-  const background = clonedNode.style?.getPropertyValue('background')
-  if (background) {
-    const cssString = await embedResources(background, null, options)
-    clonedNode.style.setProperty(
-      'background',
-      cssString,
-      clonedNode.style.getPropertyPriority('background'),
-    )
+  if (!(await embedProp('background', clonedNode, options))) {
+    await embedProp('background-image', clonedNode, options)
+  }
+  if (!(await embedProp('mask', clonedNode, options))) {
+    await embedProp('mask-image', clonedNode, options)
   }
 }
 
@@ -45,6 +60,10 @@ async function embedImageNode<T extends HTMLElement | SVGImageElement>(
     const image = clonedNode as HTMLImageElement
     if (image.decode) {
       image.decode = resolve as any
+    }
+
+    if (image.loading === 'lazy') {
+      image.loading = 'eager'
     }
 
     if (isImageElement) {